Fix gitlab ci artifacts to resolve Gitlab build failures
It seems that .gradle/* must be included between builds if build/* is
included, or Gradle will get confused and throw unusual errors on a
`publish`.

Additionally, move the custom gradle-home to `.gradle-home` to avoid
mixing of the directories, and allow easier caching of just the desired
components (downloaded artifacts and the gradle wrapper) between builds,
without bringing any build state. Additionally, makes it easier to bring
build-state between steps in a build (.gradle/*) without bringing
unwanted components that should be in the cache.